[fork] the implementation of the `docker_container` task
The goal is to be able to use Docker API version 1.19, which is all
that RedHat 6 (and the ancient build of Docker from
https://get.docker.com/ that we install) can offer.

- Tell `requirements.yml` to download [our forked
version](https://github.com/epfl-si/community.docker/tree/backport/docker_api_v1_19)
of the `community.docker` Ansible collection, which we already
[patched](https://github.com/epfl-si/community.docker/commit/5a0dd5a222ad2807780dcd86cec0006878872a64)
to use Docker API version 1.19 (spoiler: it just works)

- Bump Ansible
to version 2.10.x, so as to be able to check out the `` collection from GitHub
(instead of the Ansible Galaxy, to which we are both unable and
unwilling to publish our changes at the moment)

- Be sure to fully-qualify the `docker_container` task (singular for
now) in `ansible/tasks/prometheus.yml`, i.e.
`community.docker.docker_container:`. If we don't, some kind of
compatibility dispatch table (that we couldn't pinpoint in the source
code, in spite of a 2+-hour debugging effort) comes into play and
redirects thusly: `docker_container` →
`ansible.builtin.docker_container` →
`community.general.docker_container`. The latter (on top of being the
wrong name w.r.t. the thing we decided to fork) appears to get loaded
from Ansible's built-in `action_plugins` directory in `sys.path`, no
matter what the `ANSIBLE_COLLECTIONS_PATHS` variable is set to.

- Remove cargo-culted `community.general` line inside `requirements.yml`